requirements.txt
fastapi
uvicorn
python-multipart
openai>=1.35.13
watchdog
openpyxl
jinja2
python-dotenv
pydantic>=2.0.0
PyMuPDF
pdf2image
Pillow
pdfplumber
itsdangerous